+++ examples/opentelemetry/pyproject.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 requires-python = ">=3.10"
 dependencies = [
     "graphiti-core",
-    "kuzu>=0.11.2",
+    "real_ladybug>=0.15.1",
     "opentelemetry-api>=1.20.0",
     "opentelemetry-sdk>=1.20.0",
 ]
