+++ graphiti_core/driver/driver.py
@@ -59,8 +59,14 @@ ENTITY_EDGE_INDEX_NAME = os.environ.get('ENTITY_EDGE_INDEX_NAME', 'entity_edges'
 class GraphProvider(Enum):
     NEO4J = 'neo4j'
     FALKORDB = 'falkordb'
-    KUZU = 'kuzu'
+    LADYBUG = 'ladybug'
     NEPTUNE = 'neptune'
+
+    @classmethod
+    def _missing_(cls, value: object):
+        if isinstance(value, str) and value.lower() == 'ladybug':
+            return cls.LADYBUG
+        return None
 
 
 class GraphDriverSession(ABC):
