+++ graphiti_core/driver/ladybug/operations/entity_edge_ops.py
@@ -19,7 +19,7 @@ import logging
 from typing import Any
 
 from graphiti_core.driver.driver import GraphProvider
-from graphiti_core.driver.kuzu.operations.record_parsers import parse_kuzu_entity_edge
+from graphiti_core.driver.ladybug.operations.record_parsers import parse_ladybug_entity_edge
 from graphiti_core.driver.operations.entity_edge_ops import EntityEdgeOperations
 from graphiti_core.driver.query_executor import QueryExecutor, Transaction
 from graphiti_core.edges import EntityEdge
@@ -32,7 +32,7 @@ from graphiti_core.models.edges.edge_db_queries import (
 logger = logging.getLogger(__name__)
 
 
-class KuzuEntityEdgeOperations(EntityEdgeOperations):
+class LadybugEntityEdgeOperations(EntityEdgeOperations):
     async def save(
         self,
         executor: QueryExecutor,
@@ -55,7 +55,7 @@ class KuzuEntityEdgeOperations(EntityEdgeOperations):
             'attributes': json.dumps(edge.attributes or {}),
         }
 
-        query = get_entity_edge_save_query(GraphProvider.KUZU)
+        query = get_entity_edge_save_query(GraphProvider.LADYBUG)
         if tx is not None:
             await tx.run(query, **params)
         else:
@@ -70,7 +70,7 @@ class KuzuEntityEdgeOperations(EntityEdgeOperations):
         tx: Transaction | None = None,
         batch_size: int = 100,
     ) -> None:
-        # Kuzu doesn't support UNWIND - iterate and save individually
+        # Ladybug doesn't support UNWIND - iterate and save individually
         for edge in edges:
             await self.save(executor, edge, tx=tx)
 
@@ -115,9 +115,9 @@ class KuzuEntityEdgeOperations(EntityEdgeOperations):
         query = """
             MATCH (n:Entity)-[:RELATES_TO]->(e:RelatesToNode_ {uuid: $uuid})-[:RELATES_TO]->(m:Entity)
             RETURN
-            """ + get_entity_edge_return_query(GraphProvider.KUZU)
+            """ + get_entity_edge_return_query(GraphProvider.LADYBUG)
         records, _, _ = await executor.execute_query(query, uuid=uuid)
-        edges = [parse_kuzu_entity_edge(r) for r in records]
+        edges = [parse_ladybug_entity_edge(r) for r in records]
         if len(edges) == 0:
             raise EdgeNotFoundError(uuid)
         return edges[0]
@@ -133,9 +133,9 @@ class KuzuEntityEdgeOperations(EntityEdgeOperations):
             MATCH (n:Entity)-[:RELATES_TO]->(e:RelatesToNode_)-[:RELATES_TO]->(m:Entity)
             WHERE e.uuid IN $uuids
             RETURN
-            """ + get_entity_edge_return_query(GraphProvider.KUZU)
+            """ + get_entity_edge_return_query(GraphProvider.LADYBUG)
         records, _, _ = await executor.execute_query(query, uuids=uuids)
-        return [parse_kuzu_entity_edge(r) for r in records]
+        return [parse_ladybug_entity_edge(r) for r in records]
 
     async def get_by_group_ids(
         self,
@@ -155,7 +155,7 @@ class KuzuEntityEdgeOperations(EntityEdgeOperations):
             + """
             RETURN
             """
-            + get_entity_edge_return_query(GraphProvider.KUZU)
+            + get_entity_edge_return_query(GraphProvider.LADYBUG)
             + """
             ORDER BY e.uuid DESC
             """
@@ -167,7 +167,7 @@ class KuzuEntityEdgeOperations(EntityEdgeOperations):
             uuid=uuid_cursor,
             limit=limit,
         )
-        return [parse_kuzu_entity_edge(r) for r in records]
+        return [parse_ladybug_entity_edge(r) for r in records]
 
     async def get_between_nodes(
         self,
@@ -178,13 +178,13 @@ class KuzuEntityEdgeOperations(EntityEdgeOperations):
         query = """
             MATCH (n:Entity {uuid: $source_node_uuid})-[:RELATES_TO]->(e:RelatesToNode_)-[:RELATES_TO]->(m:Entity {uuid: $target_node_uuid})
             RETURN
-            """ + get_entity_edge_return_query(GraphProvider.KUZU)
+            """ + get_entity_edge_return_query(GraphProvider.LADYBUG)
         records, _, _ = await executor.execute_query(
             query,
             source_node_uuid=source_node_uuid,
             target_node_uuid=target_node_uuid,
         )
-        return [parse_kuzu_entity_edge(r) for r in records]
+        return [parse_ladybug_entity_edge(r) for r in records]
 
     async def get_by_node_uuid(
         self,
@@ -194,9 +194,9 @@ class KuzuEntityEdgeOperations(EntityEdgeOperations):
         query = """
             MATCH (n:Entity {uuid: $node_uuid})-[:RELATES_TO]->(e:RelatesToNode_)-[:RELATES_TO]->(m:Entity)
             RETURN
-            """ + get_entity_edge_return_query(GraphProvider.KUZU)
+            """ + get_entity_edge_return_query(GraphProvider.LADYBUG)
         records, _, _ = await executor.execute_query(query, node_uuid=node_uuid)
-        return [parse_kuzu_entity_edge(r) for r in records]
+        return [parse_ladybug_entity_edge(r) for r in records]
 
     async def load_embeddings(
         self,
