+++ graphiti_core/driver/ladybug/operations/entity_node_ops.py
@@ -14,71 +14,85 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+import json
 import logging
 from typing import Any
 
 from graphiti_core.driver.driver import GraphProvider
-from graphiti_core.driver.operations.community_node_ops import CommunityNodeOperations
+from graphiti_core.driver.ladybug.operations.record_parsers import parse_ladybug_entity_node
+from graphiti_core.driver.operations.entity_node_ops import EntityNodeOperations
 from graphiti_core.driver.query_executor import QueryExecutor, Transaction
-from graphiti_core.driver.record_parsers import community_node_from_record
 from graphiti_core.errors import NodeNotFoundError
 from graphiti_core.models.nodes.node_db_queries import (
-    COMMUNITY_NODE_RETURN,
-    get_community_node_save_query,
+    get_entity_node_return_query,
+    get_entity_node_save_query,
 )
-from graphiti_core.nodes import CommunityNode
+from graphiti_core.nodes import EntityNode
 
 logger = logging.getLogger(__name__)
 
 
-class KuzuCommunityNodeOperations(CommunityNodeOperations):
+class LadybugEntityNodeOperations(EntityNodeOperations):
     async def save(
         self,
         executor: QueryExecutor,
-        node: CommunityNode,
+        node: EntityNode,
         tx: Transaction | None = None,
     ) -> None:
-        query = get_community_node_save_query(GraphProvider.KUZU)
+        # Ladybug uses individual SET per property, attributes serialized as JSON
+        attrs_json = json.dumps(node.attributes or {})
         params: dict[str, Any] = {
             'uuid': node.uuid,
             'name': node.name,
+            'name_embedding': node.name_embedding,
             'group_id': node.group_id,
             'summary': node.summary,
-            'name_embedding': node.name_embedding,
             'created_at': node.created_at,
+            'labels': list(set(node.labels + ['Entity'])),
+            'attributes': attrs_json,
         }
+
+        query = get_entity_node_save_query(GraphProvider.LADYBUG, '')
+
         if tx is not None:
             await tx.run(query, **params)
         else:
             await executor.execute_query(query, **params)
 
-        logger.debug(f'Saved Community Node to Graph: {node.uuid}')
+        logger.debug(f'Saved Node to Graph: {node.uuid}')
 
     async def save_bulk(
         self,
         executor: QueryExecutor,
-        nodes: list[CommunityNode],
+        nodes: list[EntityNode],
         tx: Transaction | None = None,
         batch_size: int = 100,
     ) -> None:
-        # Kuzu doesn't support UNWIND - iterate and save individually
+        # Ladybug doesn't support UNWIND - iterate and save individually
         for node in nodes:
             await self.save(executor, node, tx=tx)
 
     async def delete(
         self,
         executor: QueryExecutor,
-        node: CommunityNode,
+        node: EntityNode,
         tx: Transaction | None = None,
     ) -> None:
-        query = """
-            MATCH (n:Community {uuid: $uuid})
+        # Also delete connected RelatesToNode_ intermediates
+        cleanup_query = """
+            MATCH (n:Entity {uuid: $uuid})-[:RELATES_TO]->(r:RelatesToNode_)
+            DETACH DELETE r
+        """
+        delete_query = """
+            MATCH (n:Entity {uuid: $uuid})
             DETACH DELETE n
         """
         if tx is not None:
-            await tx.run(query, uuid=node.uuid)
+            await tx.run(cleanup_query, uuid=node.uuid)
+            await tx.run(delete_query, uuid=node.uuid)
         else:
-            await executor.execute_query(query, uuid=node.uuid)
+            await executor.execute_query(cleanup_query, uuid=node.uuid)
+            await executor.execute_query(delete_query, uuid=node.uuid)
 
         logger.debug(f'Deleted Node: {node.uuid}')
 
@@ -89,14 +103,20 @@ class KuzuCommunityNodeOperations(CommunityNodeOperations):
         tx: Transaction | None = None,
         batch_size: int = 100,
     ) -> None:
-        # Kuzu doesn't support IN TRANSACTIONS OF - simple delete
+        # Clean up RelatesToNode_ intermediates first
+        cleanup_query = """
+            MATCH (n:Entity {group_id: $group_id})-[:RELATES_TO]->(r:RelatesToNode_)
+            DETACH DELETE r
+        """
         query = """
-            MATCH (n:Community {group_id: $group_id})
+            MATCH (n:Entity {group_id: $group_id})
             DETACH DELETE n
         """
         if tx is not None:
+            await tx.run(cleanup_query, group_id=group_id)
             await tx.run(query, group_id=group_id)
         else:
+            await executor.execute_query(cleanup_query, group_id=group_id)
             await executor.execute_query(query, group_id=group_id)
 
     async def delete_by_uuids(
@@ -106,31 +126,34 @@ class KuzuCommunityNodeOperations(CommunityNodeOperations):
         tx: Transaction | None = None,
         batch_size: int = 100,
     ) -> None:
-        # Kuzu doesn't support IN TRANSACTIONS OF - simple delete
+        cleanup_query = """
+            MATCH (n:Entity)-[:RELATES_TO]->(r:RelatesToNode_)
+            WHERE n.uuid IN $uuids
+            DETACH DELETE r
+        """
         query = """
-            MATCH (n:Community)
+            MATCH (n:Entity)
             WHERE n.uuid IN $uuids
             DETACH DELETE n
         """
         if tx is not None:
+            await tx.run(cleanup_query, uuids=uuids)
             await tx.run(query, uuids=uuids)
         else:
+            await executor.execute_query(cleanup_query, uuids=uuids)
             await executor.execute_query(query, uuids=uuids)
 
     async def get_by_uuid(
         self,
         executor: QueryExecutor,
         uuid: str,
-    ) -> CommunityNode:
-        query = (
-            """
-            MATCH (c:Community {uuid: $uuid})
+    ) -> EntityNode:
+        query = """
+            MATCH (n:Entity {uuid: $uuid})
             RETURN
-            """
-            + COMMUNITY_NODE_RETURN
-        )
+            """ + get_entity_node_return_query(GraphProvider.LADYBUG)
         records, _, _ = await executor.execute_query(query, uuid=uuid)
-        nodes = [community_node_from_record(r) for r in records]
+        nodes = [parse_ladybug_entity_node(r) for r in records]
         if len(nodes) == 0:
             raise NodeNotFoundError(uuid)
         return nodes[0]
@@ -139,17 +162,14 @@ class KuzuCommunityNodeOperations(CommunityNodeOperations):
         self,
         executor: QueryExecutor,
         uuids: list[str],
-    ) -> list[CommunityNode]:
-        query = (
-            """
-            MATCH (c:Community)
-            WHERE c.uuid IN $uuids
+    ) -> list[EntityNode]:
+        query = """
+            MATCH (n:Entity)
+            WHERE n.uuid IN $uuids
             RETURN
-            """
-            + COMMUNITY_NODE_RETURN
-        )
+            """ + get_entity_node_return_query(GraphProvider.LADYBUG)
         records, _, _ = await executor.execute_query(query, uuids=uuids)
-        return [community_node_from_record(r) for r in records]
+        return [parse_ladybug_entity_node(r) for r in records]
 
     async def get_by_group_ids(
         self,
@@ -157,21 +177,21 @@ class KuzuCommunityNodeOperations(CommunityNodeOperations):
         group_ids: list[str],
         limit: int | None = None,
         uuid_cursor: str | None = None,
-    ) -> list[CommunityNode]:
-        cursor_clause = 'AND c.uuid < $uuid' if uuid_cursor else ''
+    ) -> list[EntityNode]:
+        cursor_clause = 'AND n.uuid < $uuid' if uuid_cursor else ''
         limit_clause = 'LIMIT $limit' if limit is not None else ''
         query = (
             """
-            MATCH (c:Community)
-            WHERE c.group_id IN $group_ids
+            MATCH (n:Entity)
+            WHERE n.group_id IN $group_ids
             """
             + cursor_clause
             + """
             RETURN
             """
-            + COMMUNITY_NODE_RETURN
+            + get_entity_node_return_query(GraphProvider.LADYBUG)
             + """
-            ORDER BY c.uuid DESC
+            ORDER BY n.uuid DESC
             """
             + limit_clause
         )
@@ -181,18 +201,36 @@ class KuzuCommunityNodeOperations(CommunityNodeOperations):
             uuid=uuid_cursor,
             limit=limit,
         )
-        return [community_node_from_record(r) for r in records]
+        return [parse_ladybug_entity_node(r) for r in records]
 
-    async def load_name_embedding(
+    async def load_embeddings(
         self,
         executor: QueryExecutor,
-        node: CommunityNode,
+        node: EntityNode,
     ) -> None:
         query = """
-            MATCH (c:Community {uuid: $uuid})
-            RETURN c.name_embedding AS name_embedding
+            MATCH (n:Entity {uuid: $uuid})
+            RETURN n.name_embedding AS name_embedding
         """
         records, _, _ = await executor.execute_query(query, uuid=node.uuid)
         if len(records) == 0:
             raise NodeNotFoundError(node.uuid)
         node.name_embedding = records[0]['name_embedding']
+
+    async def load_embeddings_bulk(
+        self,
+        executor: QueryExecutor,
+        nodes: list[EntityNode],
+        batch_size: int = 100,
+    ) -> None:
+        uuids = [n.uuid for n in nodes]
+        query = """
+            MATCH (n:Entity)
+            WHERE n.uuid IN $uuids
+            RETURN DISTINCT n.uuid AS uuid, n.name_embedding AS name_embedding
+        """
+        records, _, _ = await executor.execute_query(query, uuids=uuids)
+        embedding_map = {r['uuid']: r['name_embedding'] for r in records}
+        for node in nodes:
+            if node.uuid in embedding_map:
+                node.name_embedding = embedding_map[node.uuid]
