+++ graphiti_core/driver/ladybug/operations/has_episode_edge_ops.py
@@ -17,104 +17,87 @@ limitations under the License.
 import logging
 from typing import Any
 
-from graphiti_core.driver.driver import GraphProvider
-from graphiti_core.driver.operations.saga_node_ops import SagaNodeOperations
+from graphiti_core.driver.operations.has_episode_edge_ops import HasEpisodeEdgeOperations
 from graphiti_core.driver.query_executor import QueryExecutor, Transaction
-from graphiti_core.errors import NodeNotFoundError
+from graphiti_core.edges import HasEpisodeEdge
+from graphiti_core.errors import EdgeNotFoundError
 from graphiti_core.helpers import parse_db_date
-from graphiti_core.models.nodes.node_db_queries import SAGA_NODE_RETURN, get_saga_node_save_query
-from graphiti_core.nodes import SagaNode
+from graphiti_core.models.edges.edge_db_queries import (
+    HAS_EPISODE_EDGE_RETURN,
+    HAS_EPISODE_EDGE_SAVE,
+)
 
 logger = logging.getLogger(__name__)
 
 
-def _saga_node_from_record(record: Any) -> SagaNode:
-    return SagaNode(
+def _has_episode_edge_from_record(record: Any) -> HasEpisodeEdge:
+    return HasEpisodeEdge(
         uuid=record['uuid'],
-        name=record['name'],
         group_id=record['group_id'],
+        source_node_uuid=record['source_node_uuid'],
+        target_node_uuid=record['target_node_uuid'],
         created_at=parse_db_date(record['created_at']),  # type: ignore[arg-type]
     )
 
 
-class KuzuSagaNodeOperations(SagaNodeOperations):
+class LadybugHasEpisodeEdgeOperations(HasEpisodeEdgeOperations):
     async def save(
         self,
         executor: QueryExecutor,
-        node: SagaNode,
+        edge: HasEpisodeEdge,
         tx: Transaction | None = None,
     ) -> None:
-        query = get_saga_node_save_query(GraphProvider.KUZU)
         params: dict[str, Any] = {
-            'uuid': node.uuid,
-            'name': node.name,
-            'group_id': node.group_id,
-            'created_at': node.created_at,
+            'saga_uuid': edge.source_node_uuid,
+            'episode_uuid': edge.target_node_uuid,
+            'uuid': edge.uuid,
+            'group_id': edge.group_id,
+            'created_at': edge.created_at,
         }
         if tx is not None:
-            await tx.run(query, **params)
+            await tx.run(HAS_EPISODE_EDGE_SAVE, **params)
         else:
-            await executor.execute_query(query, **params)
+            await executor.execute_query(HAS_EPISODE_EDGE_SAVE, **params)
 
-        logger.debug(f'Saved Saga Node to Graph: {node.uuid}')
+        logger.debug(f'Saved Edge to Graph: {edge.uuid}')
 
     async def save_bulk(
         self,
         executor: QueryExecutor,
-        nodes: list[SagaNode],
+        edges: list[HasEpisodeEdge],
         tx: Transaction | None = None,
         batch_size: int = 100,
     ) -> None:
-        # Kuzu doesn't support UNWIND - iterate and save individually
-        for node in nodes:
-            await self.save(executor, node, tx=tx)
+        for edge in edges:
+            await self.save(executor, edge, tx=tx)
 
     async def delete(
         self,
         executor: QueryExecutor,
-        node: SagaNode,
+        edge: HasEpisodeEdge,
         tx: Transaction | None = None,
     ) -> None:
         query = """
-            MATCH (n:Saga {uuid: $uuid})
-            DETACH DELETE n
+            MATCH (n:Saga)-[e:HAS_EPISODE {uuid: $uuid}]->(m:Episodic)
+            DELETE e
         """
         if tx is not None:
-            await tx.run(query, uuid=node.uuid)
+            await tx.run(query, uuid=edge.uuid)
         else:
-            await executor.execute_query(query, uuid=node.uuid)
+            await executor.execute_query(query, uuid=edge.uuid)
 
-        logger.debug(f'Deleted Node: {node.uuid}')
-
-    async def delete_by_group_id(
-        self,
-        executor: QueryExecutor,
-        group_id: str,
-        tx: Transaction | None = None,
-        batch_size: int = 100,
-    ) -> None:
-        # Kuzu doesn't support IN TRANSACTIONS OF - simple delete
-        query = """
-            MATCH (n:Saga {group_id: $group_id})
-            DETACH DELETE n
-        """
-        if tx is not None:
-            await tx.run(query, group_id=group_id)
-        else:
-            await executor.execute_query(query, group_id=group_id)
+        logger.debug(f'Deleted Edge: {edge.uuid}')
 
     async def delete_by_uuids(
         self,
         executor: QueryExecutor,
         uuids: list[str],
         tx: Transaction | None = None,
-        batch_size: int = 100,
     ) -> None:
-        # Kuzu doesn't support IN TRANSACTIONS OF - simple delete
         query = """
-            MATCH (n:Saga)
-            WHERE n.uuid IN $uuids
-            DETACH DELETE n
+            MATCH (n:Saga)-[e:HAS_EPISODE]->(m:Episodic)
+            WHERE e.uuid IN $uuids
+            DELETE e
         """
         if tx is not None:
             await tx.run(query, uuids=uuids)
@@ -125,35 +108,35 @@ class KuzuSagaNodeOperations(SagaNodeOperations):
         self,
         executor: QueryExecutor,
         uuid: str,
-    ) -> SagaNode:
+    ) -> HasEpisodeEdge:
         query = (
             """
-            MATCH (s:Saga {uuid: $uuid})
+            MATCH (n:Saga)-[e:HAS_EPISODE {uuid: $uuid}]->(m:Episodic)
             RETURN
             """
-            + SAGA_NODE_RETURN
+            + HAS_EPISODE_EDGE_RETURN
         )
         records, _, _ = await executor.execute_query(query, uuid=uuid)
-        nodes = [_saga_node_from_record(r) for r in records]
-        if len(nodes) == 0:
-            raise NodeNotFoundError(uuid)
-        return nodes[0]
+        edges = [_has_episode_edge_from_record(r) for r in records]
+        if len(edges) == 0:
+            raise EdgeNotFoundError(uuid)
+        return edges[0]
 
     async def get_by_uuids(
         self,
         executor: QueryExecutor,
         uuids: list[str],
-    ) -> list[SagaNode]:
+    ) -> list[HasEpisodeEdge]:
         query = (
             """
-            MATCH (s:Saga)
-            WHERE s.uuid IN $uuids
+            MATCH (n:Saga)-[e:HAS_EPISODE]->(m:Episodic)
+            WHERE e.uuid IN $uuids
             RETURN
             """
-            + SAGA_NODE_RETURN
+            + HAS_EPISODE_EDGE_RETURN
         )
         records, _, _ = await executor.execute_query(query, uuids=uuids)
-        return [_saga_node_from_record(r) for r in records]
+        return [_has_episode_edge_from_record(r) for r in records]
 
     async def get_by_group_ids(
         self,
@@ -161,21 +144,21 @@ class KuzuSagaNodeOperations(SagaNodeOperations):
         group_ids: list[str],
         limit: int | None = None,
         uuid_cursor: str | None = None,
-    ) -> list[SagaNode]:
-        cursor_clause = 'AND s.uuid < $uuid' if uuid_cursor else ''
+    ) -> list[HasEpisodeEdge]:
+        cursor_clause = 'AND e.uuid < $uuid' if uuid_cursor else ''
         limit_clause = 'LIMIT $limit' if limit is not None else ''
         query = (
             """
-            MATCH (s:Saga)
-            WHERE s.group_id IN $group_ids
+            MATCH (n:Saga)-[e:HAS_EPISODE]->(m:Episodic)
+            WHERE e.group_id IN $group_ids
             """
             + cursor_clause
             + """
             RETURN
             """
-            + SAGA_NODE_RETURN
+            + HAS_EPISODE_EDGE_RETURN
             + """
-            ORDER BY s.uuid DESC
+            ORDER BY e.uuid DESC
             """
             + limit_clause
         )
@@ -185,4 +168,4 @@ class KuzuSagaNodeOperations(SagaNodeOperations):
             uuid=uuid_cursor,
             limit=limit,
         )
-        return [_saga_node_from_record(r) for r in records]
+        return [_has_episode_edge_from_record(r) for r in records]
