+++ graphiti_core/models/nodes/node_db_queries.py
@@ -28,7 +28,7 @@ def get_episode_node_save_query(provider: GraphProvider) -> str:
                 entity_edges: join([x IN coalesce($entity_edges, []) | toString(x) ], '|'), created_at: $created_at, valid_at: $valid_at}
                 RETURN n.uuid AS uuid
             """
-        case GraphProvider.KUZU:
+        case GraphProvider.LADYBUG:
             return """
                 MERGE (n:Episodic {uuid: $uuid})
                 SET
@@ -69,7 +69,7 @@ def get_episode_node_save_bulk_query(provider: GraphProvider) -> str:
                 entity_edges: join([x IN coalesce(episode.entity_edges, []) | toString(x) ], '|'), created_at: episode.created_at, valid_at: episode.valid_at}
                 RETURN n.uuid AS uuid
             """
-        case GraphProvider.KUZU:
+        case GraphProvider.LADYBUG:
             return """
                 MERGE (n:Episodic {uuid: $uuid})
                 SET
@@ -136,7 +136,7 @@ def get_entity_node_save_query(provider: GraphProvider, labels: str, has_aoss: b
                 SET n.name_embedding = vecf32($entity_data.name_embedding)
                 RETURN n.uuid AS uuid
             """
-        case GraphProvider.KUZU:
+        case GraphProvider.LADYBUG:
             return """
                 MERGE (n:Entity {uuid: $uuid})
                 SET
@@ -220,7 +220,7 @@ def get_entity_node_save_bulk_query(
                     """
                 )
             return queries
-        case GraphProvider.KUZU:
+        case GraphProvider.LADYBUG:
             return """
                 MERGE (n:Entity {uuid: $uuid})
                 SET
@@ -255,7 +255,7 @@ def get_entity_node_save_bulk_query(
 
 def get_entity_node_return_query(provider: GraphProvider) -> str:
     # `name_embedding` is not returned by default and must be loaded manually using `load_name_embedding()`.
-    if provider == GraphProvider.KUZU:
+    if provider == GraphProvider.LADYBUG:
         return """
             n.uuid AS uuid,
             n.name AS name,
@@ -292,7 +292,7 @@ def get_community_node_save_query(provider: GraphProvider) -> str:
                 SET n.name_embedding = join([x IN coalesce($name_embedding, []) | toString(x) ], ",")
                 RETURN n.uuid AS uuid
             """
-        case GraphProvider.KUZU:
+        case GraphProvider.LADYBUG:
             return """
                 MERGE (n:Community {uuid: $uuid})
                 SET
@@ -333,7 +333,7 @@ COMMUNITY_NODE_RETURN_NEPTUNE = """
 
 def get_saga_node_save_query(provider: GraphProvider) -> str:
     match provider:
-        case GraphProvider.KUZU:
+        case GraphProvider.LADYBUG:
             return """
                 MERGE (n:Saga {uuid: $uuid})
                 SET
