+++ graphiti_core/nodes.py
@@ -118,7 +118,7 @@ class Node(BaseModel, ABC):
                     uuid=self.uuid,
                 )
 
-            case GraphProvider.KUZU:
+            case GraphProvider.LADYBUG:
                 for label in ['Episodic', 'Community']:
                     await driver.execute_query(
                         f"""
@@ -187,7 +187,7 @@ class Node(BaseModel, ABC):
                         batch_size=batch_size,
                     )
 
-            case GraphProvider.KUZU:
+            case GraphProvider.LADYBUG:
                 for label in ['Episodic', 'Community']:
                     await driver.execute_query(
                         f"""
@@ -243,7 +243,7 @@ class Node(BaseModel, ABC):
                         """,
                         uuids=uuids,
                     )
-            case GraphProvider.KUZU:
+            case GraphProvider.LADYBUG:
                 for label in ['Episodic', 'Community']:
                     await driver.execute_query(
                         f"""
@@ -542,7 +542,7 @@ class EntityNode(Node):
             'created_at': self.created_at,
         }
 
-        if driver.provider == GraphProvider.KUZU:
+        if driver.provider == GraphProvider.LADYBUG:
             entity_data['attributes'] = json.dumps(self.attributes)
             entity_data['labels'] = list(set(self.labels + ['Entity']))
             result = await driver.execute_query(
@@ -1010,7 +1010,7 @@ def get_episodic_node_from_record(record: Any) -> EpisodicNode:
 
 
 def get_entity_node_from_record(record: Any, provider: GraphProvider) -> EntityNode:
-    if provider == GraphProvider.KUZU:
+    if provider == GraphProvider.LADYBUG:
         attributes = json.loads(record['attributes']) if record['attributes'] else {}
     else:
         attributes = record['attributes']
