+++ pyproject.toml
@@ -28,7 +28,7 @@ Repository = "https://github.com/getzep/graphiti"
 anthropic = ["anthropic>=0.49.0"]
 groq = ["groq>=0.2.0"]
 google-genai = ["google-genai>=1.62.0"]
-kuzu = ["kuzu>=0.11.3"]
+ladybug = ["real_ladybug>=0.15.1"]
 falkordb = ["falkordb>=1.1.2,<2.0.0"]
 voyageai = ["voyageai>=0.2.3"]
 gliner2 = ["gliner2>=1.2.0; python_version>='3.11'"]
@@ -42,7 +42,7 @@ dev = [
     "anthropic>=0.49.0",
     "google-genai>=1.8.0",
     "falkordb>=1.1.2,<2.0.0",
-    "kuzu>=0.11.3",
+    "real_ladybug>=0.15.1",
     "boto3>=1.39.16",
     "opensearch-py>=3.0.0",
     "langchain-aws>=0.2.29",
